Refactor alt app to use component containers
